Implemented node server api.
