+++ app/containers/ProductSelectionPage/sagas.js
@@ -25,283 +25,17 @@ export function* productsData() {
   yield cancel(watcher);
 }
 
-export function* searchedProducts(data) {
+export function* searchedProducts(action) {
   try {
-    const repos = [
-      {
-        id: '123',
-        code: 'P121',
-        name: 'abcd',
-        type: 'Product/Bundle',
-        isBundled: true,
-        isDisableReconfiguration: true,
-        groupId: 123,
-        markup: 123,
-        quantity: {
-          value: 123,
-          isEditable: true,
-          isVisible: true,
-          dataType: 'text/select/textarea/inputSelect',
-          selectValues: [
-            {
-              id: 123,
-              value: ':List',
-              isSelected: true,
-            },
-          ],
-        },
-        discountSchedule: {
-          id: 123,
-          name: 'Diwali',
-          discountUnit: 'Percent/Amount',
-          type: 'Range/Slab',
-          tiers: [
-            {
-              id: 123,
-              name: 'tier1',
-              lowerBound: 1,
-              upperBound: 10,
-              discountpercent: 10,
-              discountamount: 123,
-            },
-          ],
-        },
-        canClone: true,
-        canSegment: false,
-        segmentData: null,
-        canReconfigure: true,
-        canShowDiscountScheduler: true,
-        listPrice: {
-          value: 123,
-          isEditable: true,
-          isVisible: true,
-          dataType: 'text/select/textarea/inputSelect',
-          selectValues: [
-            {
-              id: 123,
-              value: ':List',
-              isSelected: true,
-            },
-          ],
-        },
-        isTaxable: true,
-        additionalDiscount: {
-          value: 123,
-          isEditable: true,
-          isVisible: true,
-          dataType: 'text/select/textarea/inputSelect',
-          selectValues: [
-            {
-              id: 123,
-              value: ':List',
-              isSelected: true,
-            },
-          ],
-        },
-        netUnitPrice: 123,
-        totalPrice: 123,
-        netTotal: 123,
-        pricingMethod: {
-          values: [
-            {
-              id: 123,
-              value: ':List',
-              isSelected: true,
-            },
-            {
-              id: 123,
-              value: ':Cost',
-              isSelected: false,
-            },
-          ],
-        },
-      },
-      {
-        id: '1234',
-        code: 'P121',
-        name: 'xyz',
-        type: 'Product/Bundle',
-        isBundled: true,
-        isDisableReconfiguration: true,
-        groupId: 123,
-        markup: 123,
-        quantity: {
-          value: 123,
-          isEditable: true,
-          isVisible: true,
-          dataType: 'text/select/textarea/inputSelect',
-          selectValues: [
-            {
-              id: 123,
-              value: ':List',
-              isSelected: true,
-            },
-          ],
-        },
-        discountSchedule: {
-          id: 123,
-          name: 'Diwali',
-          discountUnit: 'Percent/Amount',
-          type: 'Range/Slab',
-          tiers: [
-            {
-              id: 123,
-              name: 'tier1',
-              lowerBound: 1,
-              upperBound: 10,
-              discountpercent: 10,
-              discountamount: 123,
-            },
-          ],
-        },
-        canClone: true,
-        canSegment: false,
-        segmentData: null,
-        canReconfigure: true,
-        canShowDiscountScheduler: true,
-        listPrice: {
-          value: 123,
-          isEditable: true,
-          isVisible: true,
-          dataType: 'text/select/textarea/inputSelect',
-          selectValues: [
-            {
-              id: 123,
-              value: ':List',
-              isSelected: true,
-            },
-          ],
-        },
-        isTaxable: true,
-        additionalDiscount: {
-          value: 123,
-          isEditable: true,
-          isVisible: true,
-          dataType: 'text/select/textarea/inputSelect',
-          selectValues: [
-            {
-              id: 123,
-              value: ':List',
-              isSelected: true,
-            },
-          ],
-        },
-        netUnitPrice: 123,
-        totalPrice: 123,
-        netTotal: 123,
-        pricingMethod: {
-          values: [
-            {
-              id: 123,
-              value: ':List',
-              isSelected: true,
-            },
-            {
-              id: 123,
-              value: ':Cost',
-              isSelected: false,
-            },
-          ],
-        },
-      },
-      {
-        id: '1235',
-        code: 'P121',
-        name: 'pqr',
-        type: 'Product/Bundle',
-        isBundled: true,
-        isDisableReconfiguration: true,
-        groupId: 123,
-        markup: 123,
-        quantity: {
-          value: 123,
-          isEditable: true,
-          isVisible: true,
-          dataType: 'text/select/textarea/inputSelect',
-          selectValues: [
-            {
-              id: 123,
-              value: ':List',
-              isSelected: true,
-            },
-          ],
-        },
-        discountSchedule: {
-          id: 123,
-          name: 'Diwali',
-          discountUnit: 'Percent/Amount',
-          type: 'Range/Slab',
-          tiers: [
-            {
-              id: 123,
-              name: 'tier1',
-              lowerBound: 1,
-              upperBound: 10,
-              discountpercent: 10,
-              discountamount: 123,
-            },
-          ],
-        },
-        canClone: true,
-        canSegment: false,
-        segmentData: null,
-        canReconfigure: true,
-        canShowDiscountScheduler: true,
-        listPrice: {
-          value: 123,
-          isEditable: true,
-          isVisible: true,
-          dataType: 'text/select/textarea/inputSelect',
-          selectValues: [
-            {
-              id: 123,
-              value: ':List',
-              isSelected: true,
-            },
-          ],
-        },
-        isTaxable: true,
-        additionalDiscount: {
-          value: 123,
-          isEditable: true,
-          isVisible: true,
-          dataType: 'text/select/textarea/inputSelect',
-          selectValues: [
-            {
-              id: 123,
-              value: ':List',
-              isSelected: true,
-            },
-          ],
-        },
-        netUnitPrice: 123,
-        totalPrice: 123,
-        netTotal: 123,
-        pricingMethod: {
-          values: [
-            {
-              id: 123,
-              value: ':List',
-              isSelected: true,
-            },
-            {
-              id: 123,
-              value: ':Cost',
-              isSelected: false,
-            },
-          ],
-        },
-      },
-    ];
-    const updatedRecord = _.filter(repos, { name: data.searchObj.searchValue });
-    if (!data.searchObj.fromSearch) {
-      yield put(searchedDataLoaded(updatedRecord));
+    const requestURL = `${`${SERVER_URL + EntityURLs.PRODUCTS}/GetProducts?GroupId=${action.searchObj.groupId}&PriceListId=${action.searchObj.priceBookId}&searchValue=${action.searchObj.searchValue}`}`;
+    const repos = yield call(request, requestURL);
+    if (!action.searchObj.fromSearch) {
+      yield put(searchedDataLoaded(repos));
     } else {
-      yield put(searchBtnDataLoaded(updatedRecord));
+      yield put(searchBtnDataLoaded(repos));
     }
-    searchBtnDataLoaded(updatedRecord);
-  } catch (err) {
-    yield put(dataLoadingError(err));
+  } catch (error) {
+    yield put(dataLoadingError(error));
   }
 }
